requirements.txt
openai
python-dotenv
chromadb 
langchain 
langgraph 
streamlit 
fastapi 
uvicorn 
scikit-learn
#langsmith